requirements.txt
telethon==1.36.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-text-splitters==0.3.0
langchain-community==0.3.0
openai==1.12.0
chromadb==0.5.0
python-dotenv==1.0.1
unstructured==0.16.0
python-docx==1.1.0
PyPDF2==3.0.1
pillow==10.3.0
python-telegram-bot==21.0
pydantic==2.9.0